requirements.txt
aiohttp==3.7.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
discord.py==1.6.0
distlib==0.3.1
filelock==3.0.12
identify==1.5.12
idna==2.10
idna-ssl==1.1.0
importlib-metadata==3.4.0
importlib-resources==5.0.0
lxml==4.6.2
multidict==5.1.0
nodeenv==1.5.0
packaging==20.8
pluggy==0.13.1
pre-commit==2.9.3
py==1.10.0
pyparsing==2.4.7
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
toml==0.10.2
tox==3.21.0
translate==3.5.0
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.3.0
yarl==1.6.3
zipp==3.4.0
